browser update

+++ src/main/java/com/woorifis/demo/model/repository/BoardRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.woorifis.demo.model.entity.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Integer>{
+public interface BoardRepository extends JpaRepository<Board, Long>{
 	
     List<Board> findByTitleContaining(String keyword);
-
+  
 }
